Adding ocean basin in the Metadata (#732)
#### What was done
- Reorganization of what's in `NetCDFToIodaConverter.h`: Moved structs
and functions to `util.h` as an attempt to make the code slightly
cleaner
- Added a simple OceanMask struct that reads in regional ocean masks
from [reccap2 ocean mask](https://reccap2-ocean.github.io/regions/). The
struct has a method to return the integer mask at a specific location.
- Fixed a bug in `trim`, related to the optional float metadata
- implementation of the new feature in the RADS converter

#### Notes
It's currently not part of the `CI`, something to think about in the
future. We could add `RECCAP2_region_masks_all_v20221025.nc` to the
tarball or stage it on hpc somewhere ...