OOPS app to replace jediinc2fv3.py (#963)
Created an OOPS app to eventually replace the jediinc2fv3.py script
which converts the JEDI increment to the FV3 increment using a new
FV3-JEDI variable change and a new fields metadata override file. The
actual replacement of jediinc2fv3.py will occur in a future Global
Workflow PR. A CTest is also included for the app.