Fix GW jjob tests for upcoming GW PR #2420 (#1041)
This PR addresses issue
[#1011](https://github.com/NOAA-EMC/GDASApp/issues/1011), related to the
failure of the "gdasatmanlvar" jjob test due to a change in the name of
the "gdasatmanlvar" run script, and, with the upcoming Global Workflow
PR [#2420](https://github.com/NOAA-EMC/global-workflow/pull/2420), the
impending failure of the "gdasatmanlfinal" jjob. This fixes the script
name in the "gdasatmanlvar" test and adds a new test for
"gdasatmanlfv3inc" which will fix the issure with "gdasatmanlfinal".

The "gdasatmanlfv3inc" and "gdasatmanlfinal" won't pass yes in GW
develop, but will after
[#2420](https://github.com/NOAA-EMC/global-workflow/pull/2420) merges GW
feature/jediinc2fv3.

This PR is just a verbatim copy of @RussTreadon-NOAA 's work, taken from
his comment
[here](https://github.com/NOAA-EMC/GDASApp/issues/1011#issuecomment-2052249333).
I re-ran the new tests, and they also passed for me in GW
feature/jediinc2fv3.