Add modulefile for Dogwood/Cactus (#1073)
While we are technically not supposed to build our own CRTM or FMS, we
are asking for forgiveness and forging ahead. This PR should allow
GDASApp to compile on WCOSS2 (Dogwood and Cactus) after a minor issue is
resolved. That issue being a compile issue in IODA that Ron is already
aware of and has a fix that works
(https://github.com/JCSDA-internal/ioda/compare/develop...bugfix/read_script_convert_fix).

---------

Co-authored-by: cory martin <cory.r.martin@clogin06.cactus.wcoss2.ncep.noaa.gov>
Co-authored-by: cory martin <cory.r.martin@clogin03.cactus.wcoss2.ncep.noaa.gov>
Co-authored-by: cory martin <cory.r.martin@clogin08.cactus.wcoss2.ncep.noaa.gov>