Use standard config files for jjob test and give output files better names (#1156)
Now that G-W
#[2661](https://github.com/NOAA-EMC/global-workflow/pull/2661) is
merged, the custom config.atmanl and config.atmensanl used by the jjob
tests can be deleted. This way, the custom files don't have to be
updated everytime the standard ones are.

I also changed the sbatch submission, so the log files for each jjob
test includes the name of the associated G-W job. This makes it a little
easier to browse the logs, without having to guess which one is which.