Add specificHumidity to ADPUPA BUFR DUMP Python API (#1246)
This PR adds specificHumidity to the ADPUPA BUFR DUMP Python API. The
following file has been changed:
~/ush/ioda/bufr2ioda/bufr2ioda_adpupa.py

The specificHumidity is calculated using the dewPoint and pressure. The
information is extracted from the prepBUFR codes and also documented.

The validation of the output obs is performed using the specificHumidity
from the prepBUFR data. For more details, please check: #1193

Co-authored-by: Praveen Singh <pkumar@hercules-login-4.hpc.msstate.edu>