Build GDASApp outside of global-workflow (#1306)
Tests in `tests/gw-ci` require global-workflow. As such, GDASApp can not
build outside of g-w. This PR places gw-ci tests under the flag
`RUN_GW_CI`. gw-ci tests are built when `RUN_GI_CW` is set to `ON`. The
default is `OFF`. With this change GDASApp can be built without g-w.

Resolves #1305