Add safety feature for testing algo changes (#35)
When the JCB tests are being run (from anywhere) the code will clone the
main JCB repo and when doing so it will look for the branch name of the
PR. Then it will run a script to clone all the clients that are
registered for testing as well as the algorithms repo. This script will
use the branch name of the main JCB repo and if found for any clients
clone that branch.

Now imagine this scenario: someone is making changes to their client
(say gdas) and simultaneously making changes to jcb-algorithms. They
might realize that they need to change another client to accommodate the
algorithm changes so they make a branch with the same name in (say)
rrfs. However if they do not have that branch in the main jcb repo there
wouldn't exist a mechanism to check out the correct branch of
jcb-algorithms and the other clients during the integration tests. We
could put that kind of logic in the GitHub script but that would be
complicated and have to propagate to every jcb repo.

The change in this code forces the user to make a branch of the main jcb
repo if they have created the same branch in the jcb-algorithms branch
and thus provide all the mechanisms for the scripts to check everything
out properly.

The algorithms should really be part of JCB since they are generic and
it would avoid the need to do this. But because of EE2 compliance we
can't have YAMLs with source code, which is why they are in a separate
repo.

Co-authored-by: danholdaway <danholdaway@users.noreply.github.com>