Switch Aerosol DA to use JCB and Jedi class (#3125)
This PR switches to using JCB for aerosol DA YAML generation + makes use
of the Jedi class refactoring already done by @DavidNew-NOAA for the
atmospheric JEDI-based DA.

Resolves #3035

Note this PR is in draft for 2 reasons. 1) Need associated GDASApp PR to
be opened and merged first and 2) requires some fix file updates. This
will be coordinated with an issue and linked to this PR once
@andytangborn is satisfied with a first-attempt at the fix files.
---------

Co-authored-by: RussTreadon-NOAA <Russ.Treadon@noaa.gov>
Co-authored-by: DavidNew-NOAA <david.new@noaa.gov>