update soil variables to be consistent with fv3-jedi changes (#56)
When the GDASApp `sorc/fv3-jedi` hash is updated to
[97e10d5](https://github.com/JCSDA-internal/fv3-jedi/commit/97e10d5addfe75debab7a6b5604e7a7443f00333)
or a hash after this commit, it is necessary, at a minimum, to update
soil variables in the following templates
```
model/atmosphere/atmosphere_background.yaml.j2
model/atmosphere/atmosphere_background_ensemble.yaml.j2
```
This PR is opened to update the soil variables in the above templates.

Resolves #49

**NOTE**: GDASApp PR
[#1377](https://github.com/NOAA-EMC/GDASApp/pull/1377) requires this PR
be merged into jcb-gdas `develop`. Once this is done, the
`parm/jcb-gdas` hash in
[#1377](https://github.com/NOAA-EMC/GDASApp/pull/1377) should be updated
to point at the new jcb-gdas `develop`.