Add options to inflate the ensemble while recentering (#1501)
# Description

Adds code to the ensemble recentering to optionally inflate/deflate the
ensemble variance by either a scalar (same for all variables) or a field
from the file (has to be specified for all variables).

This is not exercised yet, since we don't know the parameters of
inflation/deflation yet, but I tested with both options to check whether
it behaves as expected.

Note: in the future refactoring we should switch to oops::Inflation that
provides several different inflation options.

# Issues

Related to #1496 (doesn't fix yet since we need to decide what
parameters to use this with)

# Automated CI tests to run in Global Workflow
<!-- Which Global Workflow CI tests are required to adequately test this
PR? -->
- [ ] atm_jjob <!-- JEDI atm single cycle DA !-->
- [ ] C96C48_ufs_hybatmDA <!-- JEDI atm cycled DA !-->
- [ ] C96C48_hybatmaerosnowDA  <!-- JEDI aero/snow cycled DA !-->
- [ ] C48mx500_3DVarAOWCDA <!-- JEDI low-res marine 3DVar cycled DA !-->
- [x] C48mx500_hybAOWCDA <!-- JEDI marine hybrid envar cycled DA !-->
- [ ] C96C48_hybatmDA <!-- GSI atm cycled DA !-->

---------

Co-authored-by: Anna Shlyaeva <anna.v.shlyaeva@noaa.gov>
Co-authored-by: Guillaume Vernieres <guillaume.vernieres@noaa.gov>