JEDI-based ensemble recentering and analysis calculation and native grid increments (#72)
This PR is a companion to GDASApp PR
[#1488](https://github.com/NOAA-EMC/GDASApp/pull/1488) and Global
Workflow PR
[#3312](https://github.com/NOAA-EMC/global-workflow/pull/3312).

It creates YAMLs for the JEDI-based ensemble recentering application and
the `convertstate` applications used for analysis calculation. It also
alters certain YAMLs to implement native-grid increments in the Global
Workflow.

Some miscellaneous changes to the naming convention of increments and
backgrounds in run directories are also implemented for overall
consistency and simplicity across JEDI-based Global Workflow jobs.

---------

Co-authored-by: danholdaway <danholdaway@users.noreply.github.com>
Co-authored-by: Jiarui Dong <Jiarui.Dong@noaa.gov>
Co-authored-by: Cory Martin <cory.r.martin@noaa.gov>