Add precipitable clouds in microwave all-sky assimilations (updated) (#74)
This is an update version of the PR
https://github.com/NOAA-EMC/jcb-gdas/pull/63 which resolves the issue
https://github.com/NOAA-EMC/GDASApp/issues/1333 .
The main update is that calculations of cloud fraction and cloud
particle radius by the Thompson microphysics scheme are added.

Here is the summary of other added configurations in the original PR
#63:
Add rain, snow, and graupel clouds for CRTM inputs.
Add a prior filter to zero clouds over non-water surfaces.
Add an option of clouds in the Hydrometeo check.
Add gross error check threshold for precipition impacted channels.
Add cold-air-outbreak filters.
Read cloud fraction from GeoVals.
Render list of not-bias corrected channels from JCB tables.
Add configurations for atms_n21.

It depends on two UFO PRs
https://github.com/JCSDA-internal/ufo/pull/3600
https://github.com/JCSDA-internal/ufo/pull/3573

It needs **the develop branch in
"global-workflow/sorc/gdas.cd/sorc/jcb"** in order to rend non-bias
corrected channels.

These UFO PRs are being revised. I'll inform you when it is ready to be
merged.

---------

Co-authored-by: Cory Martin <cory.r.martin@noaa.gov>