Bugfix for marine qc filters (#105)
This bugfix PR addresses an ordering issue in the post filters of
sst_generic.yaml.j2 and adds "observed variables" that are missing in
several of the in situ observation yamls.

The missing "observed variables" in bathy, altkob, trkob and
trkob_salinity causes the gdas_prepoceanobs task to fail. Using the
background check at the end of the post filters in sst_generic.yaml.j2
cause the gdas_marineanlvar task to fail.
 
These bugfix changes were successfully tested on Hera.