Add initial configures enable for cycle dependent reject lists (#86)
This PR adds the initial configures enable for using the different
reject lists for the different cycles. The purpose of this PR is to use
the JCB conventional chronicle as separate reject lists to remove the
bad snow depth reports in the 06/12/18 cycles in China. This PR adds
eleven stations from China in the reject list at t00z.

This PR also removes the stations in China from reject list on October
1, 2022.

This PR contributes https://github.com/NOAA-EMC/GDASApp/pull/1549