Use UFS MOM6 and Icepack (#1659)
# Description

This PR modifies `build.sh` so that if `$WORKFLOW_BUILD == 'YES'`,
symbolic links in SOCA are created to the MOM6 and Icepack submodules in
in UFS. This is part of a broader effort of forking SOCA for the purpose
of EE2 compliance.

I've tested `C48mx500_3DVarAOWCDA` successfully on Hera with the
workflow build.

# Companion PRs

SOCA PR [#1151](https://github.com/JCSDA-internal/soca/pull/1151)

# Automated CI tests to run in Global Workflow
<!-- Which Global Workflow CI tests are required to adequately test this
PR? -->
- [ ] atm_jjob <!-- JEDI atm single cycle DA !-->
- [ ] C96C48_ufs_hybatmDA <!-- JEDI atm cycled DA !-->
- [ ] C96C48_hybatmaerosnowDA  <!-- JEDI aero/snow cycled DA !-->
- [x] C48mx500_3DVarAOWCDA <!-- JEDI low-res marine 3DVar cycled DA !-->
- [ ] C48mx500_hybAOWCDA <!-- JEDI marine hybrid envar cycled DA !-->
- [ ] C96C48_hybatmDA <!-- GSI atm cycled DA !-->

---------

Co-authored-by: Cory Martin <cory.r.martin@noaa.gov>