Relocate config templates to `dev/` space in prep for EE2 (#3684)
This PR:
- relocates `parm/config/gfs|gefs|sfs` into
`dev/parm/config/gfs|gefs|sfs`
- updates `setup_expt.py` so that the `configdir` option defaults to the
new location
- updates the included yamls to the relocated path
- updates the documentation to reflect this change
- updates affected configs to resolve shellcheck errors

In addition to the above, this PR:
- updates the yamls in the sfs to include the defaults provided in
`dev/parm/config/sfs/yaml/defaults.yaml`. Previously, they were
duplicated in the sfs test yaml
- relocates a `anlstat` yaml that was misplaced in
`parm/config/gfs/yaml`. Its correct place is in `parm/gdas`

When the release is created for GFS (or any other system), `parm/config`
will contain along with `gfs`, folders for `gefs` and `sfs`. This is
`dead code` as far as operations of the GFS is concerned.

---------

Co-authored-by: David Huber <69919478+DavidHuber-NOAA@users.noreply.github.com>