Revise IMS ioda-stats YAML and add other snow stats YAMLs (#123)
This PR directly solves Issue:
https://github.com/NOAA-EMC/jcb-gdas/issues/122

Edits made in this PR include:
1) fix a bug in the current IMS ioda-stats YAML since diag_ims file uses
longitude from 0 to 360, instead of -180 to 180.
2) add `snocvr_snow.yaml` and  `sfcsno.yaml` for outputing ioda-stats.
3) add other domains of interest to all three snow-related YAMLs per
discussions on 05/15/2025 land team meeting.

No significant changes of the output are expected as a result of this
PR. Note: the bin size of the ims_snow.yaml is set to 2 deg for now due
to run time consideration. 1 deg bin takes a much longer time for
ioda-stats.exe to process diag_ims file at a relatively fine resolution
(e.g., C384).

A G-W CI test of `C96C48mx500_S2SW_cyc_gfs` was conducted: the
concatenated `snow.yaml` looks OK. A follow-up PR will be submitted to
G-W to make sure snocvr stats (and snocvr.yaml) and sfcsno stats (and
sfcsno.yaml) can be output as well.