rename "snocvr_snow" to "snocvr" (#125)
This PR directly addresses the Issue:
https://github.com/NOAA-EMC/jcb-gdas/issues/124.

In all, (outdated) "snocvr_snow" = (now) "madis_snow" for use for
2021-2023 observations. "snocvr" is used to denote 2023 to current
snocvr observations. Based on conversations with Jiarui, several draft
PRs in G-W and GDAS will be submitted to handle the switch among
different snocvr names automatically.

Again, thanks for discussions with Jiarui Dong @jiaruidong2017 for the
clarification of the issue!