Refactor marine analysis Task class (#3642)
This PR refactors the marine analysis task class to be in line with the
standard code structure of other JEDI domains (atmosphere, aerosol,
snow). This includes:
1. Using the `JEDI` class infrastructure to initialize and run all JEDI
applications used in the marine analysis task.
2. Ensuring that the input YAML for all JEDI applications are rendered
with JCB.
3. Using `task_config` to render the JCB base YAML, rather than a custom
dictionary.
4. Ensuring that the `task_config` member dictionary of the marine
analysis `Task` class is only modified at initialization by the class
constructor.
5. Ensuring that all file staging and finalizing is carried out with
YAML files via the wxflow FileHandler class.

Additionally, a method has been added to the `Jedi` class which goes
through the YAML it generates in the `initialize()` method and removes
any observations with missing input files. This was previously done
manually in the marine analysis code, but I've absorbed it into the
`Jedi` class because it's useful/required for all domains.

Resolves #2931 

---------

Co-authored-by: Rahul Mahajan <aerorahul@users.noreply.github.com>