Add qc filtering for lon/lat (#1713)
# Description

<!-- Enter PR description here. -->
This pr adds qc for latitude and longitude to all converters.
the function clean_lat_lon checks that there are no garbage values,
and that longitude is all in [-180, 180] or all in [0, 360]
and latitude is all in [-90, 90]
Until now there was no need to qc lon and lat, but the dbuoyb converter
is picking up the entire surface temperature field from the file,
and some values of lat/lon were garbage

# Companion PRs

<!-- Enter links to any companion PRs here. -->

# Issues

<!-- Enter any issues referenced or resolved by this PR here. Use
keywords "Resolves" or "Refs".
Resolves #1234
Refs #4321
Refs NOAA-EMC/repo#5678
-->

# Automated CI tests to run in Global Workflow
<!-- Which Global Workflow CI tests are required to adequately test this
PR? -->
- [ ] atm_jjob <!-- JEDI atm single cycle DA !-->
- [ ] C96C48_ufs_hybatmDA <!-- JEDI atm cycled DA !-->
- [ ] C96C48_hybatmaerosnowDA  <!-- JEDI aero/snow cycled DA !-->
- [x] C48mx500_3DVarAOWCDA <!-- JEDI low-res marine 3DVar cycled DA !-->
- [ ] C48mx500_hybAOWCDA <!-- JEDI marine hybrid envar cycled DA !-->
- [ ] C96C48_hybatmDA <!-- GSI atm cycled DA !-->

---------

Co-authored-by: Guillaume Vernieres <guillaume.vernieres@noaa.gov>