Update madis_snow YAML and add reject list (#132)
The PR directly solves issue:
https://github.com/NOAA-EMC/jcb-gdas/issues/131

The intent is to make the madis_snow yaml consistent with other (current
as of 05/30/2025) yamls. A reject list is also added to include 77 bad
madis_snow stations identified between 2021 and 2023.

G-W tests are conducted to make sure the newly-added madis_snow reject
list is read by JEDI.