remove C96C48_hybatmaerosnowDA from g-w CI and PR template (#1736)
# Description

This PR removes `C96C48_hybatmaerosnowDA` from the GDASApp build when
`WORKFLOW_TESTS="ON"`. It also removes `C96C48_hybatmaerosnowDA` from
the list of g-w CI to run when creating GDASApp PRs. It does not remove
all references to `C96C48_hybatmaerosnowDA` from GDASApp. This can be
done as part of GDASApp issues #1734 or #1735.

This simplified PR is opened to allow the daily stable-nightly job to
continue running.

See GDASApp issue #1733 for additional information.



# Companion PRs

none

# Issues

Resolves #1733


# Automated CI tests to run in Global Workflow
<!-- Which Global Workflow CI tests are required to adequately test this
PR? -->
- [ ] atm_jjob <!-- JEDI atm single cycle DA !-->
- [ ] C96C48_ufs_hybatmDA <!-- JEDI atm cycled DA !-->
- [ ] C48mx500_3DVarAOWCDA <!-- JEDI low-res marine 3DVar cycled DA !-->
- [ ] C48mx500_hybAOWCDA <!-- JEDI marine hybrid envar cycled DA !-->
- [ ] C96C48_hybatmDA <!-- GSI atm cycled DA !-->