Add aircraft yaml j2.  (#103)
This PR adds the aircraft profiles prepbufr yaml (phase one) to the
repository.

This solves issue [JEDI T20
#116](https://github.com/issues/assigned?issue=NOAA-EMC%7CJEDI-T2O%7C116)
This should be merged at the same time as GDASApp
[PR#1597](https://github.com/NOAA-EMC/GDASApp/pull/1597/files)

Many of the paths are the ones I used to test the yaml. I added what I
believe to be the jcb paths though they will likely need work.

---------

Co-authored-by: Nicholas Esposito <esposito@hercules-login-2.hpc.msstate.edu>
Co-authored-by: Nicholas Esposito <esposito@hercules-login-4.hpc.msstate.edu>
Co-authored-by: Nicholas Esposito <esposito@hercules-login-3.hpc.msstate.edu>