Use thinning for ice obs in LETKF (#1814)
# Description

Use thinning for AMSR2 ice observations in LETKF (both makes sense
scientifically, and brings down the runtime to ~8min on hera on 480 MPI
tasks).

Since thinning is done with "reduce obs space" option, we need to run it
only once. Hence the switch to using linear observer here: in this case
filters are only run on ensemble mean.

# Companion PRs

https://github.com/NOAA-EMC/jcb-gdas/pull/153

# Issues

Resolves https://github.com/NOAA-EMC/GDASApp/issues/1542

# Automated CI tests to run in Global Workflow
<!-- Which Global Workflow CI tests are required to adequately test this
PR? -->
- [ ] atm_jjob <!-- JEDI atm single cycle DA !-->
- [ ] C96C48_ufs_hybatmDA <!-- JEDI atm cycled DA !-->
- [x] C96C48_hybatmsnowDA <!-- JEDI snow cycled DA !-->
- [ ] C96_gcafs_cycled <!-- JEDI aerosol cycled DA !-->
- [x] C48mx500_3DVarAOWCDA <!-- JEDI low-res marine 3DVar cycled DA !-->
- [x] C48mx500_hybAOWCDA <!-- JEDI marine hybrid envar cycled DA !-->
- [ ] C96C48_hybatmDA <!-- GSI atm cycled DA !-->

---------

Co-authored-by: Anna Shlyaeva <anna.v.shlyaeva@noaa.gov>