Faster marine ensemble recentering (short-term solution) (#1829)
# Description

Short-term solution to make ensemble recentering faster:
- only save recentering increment once since it's the same for all
members; copy the same increment to all ensemble members outside of jedi
app
- don't make an extra copy of CICE restarts
- remove inflation code since it's not working correctly anyway

Tested with 0.25 degree run. Ran 1.5 cycles, including the recentering
and the following ensemble forecast. Confirmed that the files generated
in COMROOT by the marine recentering task using this branch are the same
as using gdasapp as of last Thursday.

Runtime of SocaEnsHandler using 30 members, on hera with intel: before
603s, after: 232s

# Companion PRs

https://github.com/NOAA-EMC/jcb-gdas/pull/154

# Issues

Short-term solution for #1813

# Automated CI tests to run in Global Workflow
<!-- Which Global Workflow CI tests are required to adequately test this
PR? -->
- [ ] atm_jjob <!-- JEDI atm single cycle DA !-->
- [ ] C96C48_ufs_hybatmDA <!-- JEDI atm cycled DA !-->
- [x] C96C48_hybatmsnowDA <!-- JEDI snow cycled DA !-->
- [ ] C96_gcafs_cycled <!-- JEDI aerosol cycled DA !-->
- [ ] C48mx500_3DVarAOWCDA <!-- JEDI low-res marine 3DVar cycled DA !-->
- [x] C48mx500_hybAOWCDA <!-- JEDI marine hybrid envar cycled DA !-->
- [ ] C96C48_hybatmDA <!-- GSI atm cycled DA !-->

Co-authored-by: Anna Shlyaeva <anna.v.shlyaeva@noaa.gov>