Add oro file copy to ensure backwards compatibility for land SCF code (#1870)
# Description

A previous PR, preparing for an upcoming workflow change, broke
backwards compatibility with the existing IMS SCF preprocessor code.
This PR allows the workflow to copy the file needed twice, once to the
destination the old code expects and once to the destination for the new
code.

# Companion PRs

None

# Issues

None created

# Automated CI tests to run in Global Workflow
<!-- Which Global Workflow CI tests are required to adequately test this
PR? -->
- [ ] atm_jjob <!-- JEDI atm single cycle DA !-->
- [ ] C96C48_ufs_hybatmDA <!-- JEDI atm cycled DA !-->
- [x] C96C48_hybatmsnowDA <!-- JEDI snow cycled DA !-->
- [ ] C96_gcafs_cycled <!-- JEDI aerosol cycled DA !-->
- [ ] C48mx500_3DVarAOWCDA <!-- JEDI low-res marine 3DVar cycled DA !-->
- [ ] C48mx500_hybAOWCDA <!-- JEDI marine hybrid envar cycled DA !-->
- [ ] C96C48_hybatmDA <!-- GSI atm cycled DA !-->