Make IMS C++ preprocessing code use real MPI (#1908)
# Description

The initial implementation of the C++ SCF preprocessor was developed in
a way that cut corners and only worked for one MPI task per FV3 tile.
This is okay for low resolution, but does not work for C1152 (GFSv17
target). This PR introduces real MPI capabilities by leveraging the
magic of Atlas. Note that further improvements can still occur,
particularly in the input of the IMS dataset being done N number of
times instead of properly parallelized.

I have confirmed that the output looks the same for 96 and 24 tasks at
C1152, and appears to be the same as the old code at C96 when using 24
and 96 tasks.

For example:
<img width="1500" height="1000" alt="image"
src="https://github.com/user-attachments/assets/4c9c32ae-06de-497b-8141-3ee79ab4a2f9"
/>


# Companion PRs

[JCB-GDAS](https://github.com/NOAA-EMC/jcb-gdas/pull/179)

# Issues

Closes #1899 

# Automated CI tests to run in Global Workflow
<!-- Which Global Workflow CI tests are required to adequately test this
PR? -->
- [ ] atm_jjob <!-- JEDI atm single cycle DA !-->
- [ ] C96C48_ufs_hybatmDA <!-- JEDI atm cycled DA !-->
- [ ] C96C48_hybatmsnowDA <!-- JEDI snow cycled DA !-->
- [ ] C96_gcafs_cycled <!-- JEDI aerosol cycled DA !-->
- [ ] C48mx500_3DVarAOWCDA <!-- JEDI low-res marine 3DVar cycled DA !-->
- [ ] C48mx500_hybAOWCDA <!-- JEDI marine hybrid envar cycled DA !-->
- [ ] C96C48_hybatmDA <!-- GSI atm cycled DA !-->