Bring back thinning for ice obs with LETKF (#198)
Brings back thinning for ice obs with LETKF.
This option requires `use linear observer: true` in the LETKF, because
the observations are thinned with option `reduce obs space`. When obs
space is reduced, it's important to only do it once. `use linear
observer: true` ensures that the QC filters are only run once, on the
ensemble mean. When `use linear observer` is set to `false` (by
default), QC filters are run on each ensemble member, which results in
removing observations from the obs space multiple times and subsequent
segfaults in the LETKF when the various vector sizes don't match.

Goes with https://github.com/NOAA-EMC/jcb-algorithms/pull/16 and a
GDASApp PR.

Co-authored-by: shlyaeva <anna.v.shlyaeva@noaa.gov>
Co-authored-by: DavidNew-NOAA <134300700+DavidNew-NOAA@users.noreply.github.com>